% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{aws.comprehend-package}
\alias{aws.comprehend-package}
\alias{aws.comprehend}
\title{aws.comprehend}
\description{
AWS Comprehend Client Package
}
\details{
Client for AWS Comprehend (\url{https://aws.amazon.com/comprehend}0, a cloud natural language processing service that can perform a number of quantitative text analyses, including language detection, sentiment analysis, and feature extraction.
}
\seealso{
\code{\link{detect_language}}, \code{\link{detect_sentiment}}, \code{\link{detect_entities}}, \code{\link{detect_phrases}}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
