% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_phrases.R
\name{detect_phrases}
\alias{detect_phrases}
\title{Detect key phrases}
\usage{
detect_phrases(text, language = "en", ...)
}
\arguments{
\item{text}{A character string containing a text to analyze, or a character vector to perform analysis separately for each element.}

\item{language}{A character string containing a two-letter language code. Currently \dQuote{en} and \dQuote{es} are supported.}

\item{\dots}{Additional arguments passed to \code{\link{comprehendHTTP}}.}
}
\value{
A data frame
}
\description{
Detect key phrases in a source text
}
\examples{
\dontrun{
  # simple example
  detect_phrases("Amazon provides web services. Jeff is their leader.")
  
  txt <-c("Amazon provides web services.",
          "Jeff is their leader.")
  detect_phrases(txt)
}
}
