% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_sentiment.R
\name{detect_sentiment}
\alias{detect_sentiment}
\title{Detect sentiment in a source text}
\usage{
detect_sentiment(text, language = "en", ...)
}
\arguments{
\item{text}{A character string containing a text to sentiment analyze, or a character vector to perform analysis separately for each element.}

\item{language}{A character string containing a two-letter language code. Currently \dQuote{en} and \dQuote{es} are supported.}

\item{\dots}{Additional arguments passed to \code{\link{comprehendHTTP}}.}
}
\value{
A data frame
}
\description{
Detect sentiment in a source text
}
\examples{
\dontrun{
  # simple example
  detect_sentiment("I have never been happier. This is the best day ever.")

  txt <-c("I have never been happier. This is the best day ever.",
          "I have always been happier. This is the worst day ever.")
  detect_sentiment(txt)
}
}
