% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_syntax.R
\name{detect_syntax}
\alias{detect_syntax}
\title{Detect syntax in a source text}
\usage{
detect_syntax(text, language = "en", ...)
}
\arguments{
\item{text}{A character string containing a text to syntax analyze, or a character vector to perform analysis separately for each element.}

\item{language}{A character string containing a two-letter language code. Currently \dQuote{en} and \dQuote{es} are supported.}

\item{\dots}{Additional arguments passed to \code{\link{comprehendHTTP}}.}
}
\value{
A data frame
}
\description{
Detect syntax in a source text
}
\examples{
\dontrun{
  # simple example
  detect_syntax("The quick brown fox jumps over the lazy dog.")

  txt <-c("The quick brown fox jumps over the lazy dog.",
          "I have never been happier. This is the best day ever.")
  detect_syntax(txt)
}
}
