% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_and_index.R
\name{flatten}
\alias{flatten}
\title{Flatten embedded data.frames (1 level max)}
\usage{
flatten(df)
}
\arguments{
\item{df}{data.frame to flatten}
}
\description{
Flatten embedded data.frames (1 level max)
}
