% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_icd10cm.R
\name{infer_icd10cm}
\alias{infer_icd10cm}
\title{links entities to normalized concept identifiers}
\usage{
infer_icd10cm(text, ...)
}
\arguments{
\item{text}{A character string containing a text to entities analyze, or a character vector to perform analysis separately for each element.}

\item{\dots}{Additional arguments passed to \code{\link{comprehendHTTP}}.}
}
\value{
A data frame
}
\description{
detects medical conditions as entities listed in a patient record and links those entities to normalized concept identifiers 
in the ICD-10-CM knowledge base from the Centers for Disease Control.
}
\examples{
\dontrun{
  # simple example
  infer_icd10cm("Mrs. Smith has a history of diabetes mellitus.")
  
  txt <-c("Mrs. Smith has a history",
          "of diabetes mellitus")
  infer_icd10cm(txt)
}
}
