% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_rxnorm.R
\name{infer_rxnorm}
\alias{infer_rxnorm}
\title{identify medications that are listed in a patient record as entities}
\usage{
infer_rxnorm(text, ...)
}
\arguments{
\item{text}{A character string containing a text to entities analyze, or a character vector to perform analysis separately for each element.}

\item{\dots}{Additional arguments passed to \code{\link{comprehendHTTP}}.}
}
\value{
A data frame
}
\description{
identify medications that are listed in a patient record as entities
}
\examples{
\dontrun{
  # simple example
  infer_rxnrom("Mrs. Smith is currently taking ibuprofen daily.")
  
  txt <-c("Mrs. Smith is currently taking",
          "ibuprofen daily")
  infer_rxnorm(txt)
}
}
