% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_snowmed_ct.R
\name{infer_snowmed_ct}
\alias{infer_snowmed_ct}
\title{Identify clinical terms in a source medical text}
\usage{
infer_snowmed_ct(text, language = "en", ...)
}
\arguments{
\item{text}{A character string containing a text to entities analyze, or a character vector to perform analysis separately for each element.}

\item{language}{A character string containing a two-letter language code. Currently only \dQuote{en} is supported.}

\item{\dots}{Additional arguments passed to \code{\link{comprehendHTTP}}.}
}
\value{
A data frame
}
\description{
Identify clinical terms in a source medical text and link them to Systematized Nomenclature of Medicine, Clinical Terms (SNOMED CT) codes
}
\examples{
\dontrun{
  # simple example
  infer_snowmed_ct("BHEENT : Boggy inferior turbinates. No oropharyngeal lesion.")

  txt <-c("BHEENT : Boggy inferior turbinates.",
          "No oropharyngeal lesion.")
  infer_snowmed_ct(txt)
}
}
\references{
\href{https://docs.aws.amazon.com/comprehend-medical/latest/dev/ontology-linking-snomed.html}{AWS Comprehend Medical Developer Guide}
}
