% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_rxnorm.R
\name{infer_rxnorm}
\alias{infer_rxnorm}
\title{Identify medications in a source medical text}
\usage{
infer_rxnorm(text, language = "en", ...)
}
\arguments{
\item{text}{A character string containing a text to entities analyze, or a character vector to perform analysis separately for each element.}

\item{language}{A character string containing a two-letter language code. Currently only \dQuote{en} is supported.}

\item{\dots}{Additional arguments passed to \code{\link{comprehendHTTP}}.}
}
\value{
A data frame
}
\description{
Identify medications in a source medical text and link them to RxCUI codes
}
\examples{
\dontrun{
  # simple example
  infer_rxnorm("fluoride topical ( fluoride 1.1 \% topical gel ) 1 application Topically daily. Patient is not on warfarin.")

  txt <-c("fluoride topical ( fluoride 1.1 \% topical gel ) 1 application Topically daily.",
          "Patient is not on warfarin.")
  infer_rxnorm(txt)
}
}
\references{
\href{https://docs.aws.amazon.com/comprehend-medical/latest/dev/ontology-RxNorm.html}{AWS Comprehend Medical Developer Guide}
}
